moved misc and test scripts
