+++ task_manager/scripts/misc/egsr_cut.py
