+++ task_manager/scripts/test/test_vision_manager.py
